refactor(runners): extract opencode turn runner to runners/opencode.rs (Phase 2d)
run_opencode_turn (~2000 lines) moves verbatim. The SSE/JSON parsing
helpers, session storage readers, auth-state probing, and provider
injection it shares with the goal/control arm stay in mission_runner
behind pub(crate) (structs: OpencodeSseState, OpencodeSseParseResult,
OpenCodeAuthState, StoredOpenCodeMessage). No behavior change.
