+++ src/api/runners/mod.rs
@@ -9,3 +9,4 @@ pub(crate) mod codex;
 pub(crate) mod errors;
 pub(crate) mod gemini;
 pub(crate) mod grok;
+pub(crate) mod opencode;
